+++ GeoNorge/AT.Common.GeoNorge.Publish/DependencyInjection/DependencyInjectionExtensions.cs
@@ -20,10 +20,14 @@ public record GeoNorgeConfig
 /// </summary>
 public static class DependencyInjectionExtensions
 {
-    internal const string ClientKey = "GeoNorgeClient";
+    internal const string GeoNorgeClientKey = "GeoNorgeClient";
 
     /// <summary>
-    /// Register GeoNorge services in the provided <see cref="IServiceCollection"/>.
+    /// Register GeoNorge services in the provided <see cref="IServiceCollection"/>:
+    /// <br/>
+    /// - <see cref="IAddressSearch"/> for address search functionality.
+    /// <br/>
+    /// - <see cref="IFylkeKommuneApi"/> for fylke and kommune information.
     /// </summary>
     /// <param name="services"></param>
     /// <param name="geoNorgeConfig"></param>
@@ -37,7 +41,7 @@ public static class DependencyInjectionExtensions
 
         services
             .AddHttpClient(
-                ClientKey,
+                GeoNorgeClientKey,
                 client =>
                 {
                     client.BaseAddress = new Uri(geoNorgeConfig.BaseUrl);
@@ -45,7 +49,8 @@ public static class DependencyInjectionExtensions
             )
             .AddStandardResilienceHandler();
 
-        services.AddSingleton<IGeoNorge, GeoNorgeClient>();
+        services.AddSingleton<IAddressSearch, AddressSearchClient>();
+        services.AddSingleton<IFylkeKommuneApi, FylkeKommuneClient>();
 
         return services;
     }