+++ GeoNorge/AT.Common.GeoNorge.Test/Integration/Setup/GeoNorgeTestFixture.cs
@@ -1,10 +1,10 @@
 using Arbeidstilsynet.Common.GeoNorge.DependencyInjection;
-using AT.Common.GeoNorge.Test.Extensions;
+using Arbeidstilsynet.Common.TestExtensions.Extensions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using WireMock.Logging;
+using WireMock.Admin.Mappings;
+using WireMock.Net.OpenApiParser.Settings;
 using WireMock.Server;
-using WireMock.Settings;
 using Xunit.Microsoft.DependencyInjection;
 using Xunit.Microsoft.DependencyInjection.Abstracts;
 
@@ -16,21 +16,14 @@ public class GeoNorgeTestFixture : TestBedFixture
 
     public GeoNorgeTestFixture()
     {
-        _server = WireMockServer.Start(
-            new WireMockServerSettings { Logger = new WireMockConsoleLogger() }
-        );
-
-        using var fileStream = File.Open(
-            "Integration/TestData/openapi.json",
-            FileMode.Open,
-            FileAccess.Read
-        );
-
-        _server.AddOpenApiMappings(
-            fileStream,
-            m =>
+        _server = WireMockServer.Start();
+        _server.AddMappings("Integration/TestData/adresser-openapi.json");
+        _server.AddMappings(
+            "Integration/TestData/kommuner-openapi.json",
+            KommuneFylkeMappingVisitor,
+            new WireMockOpenApiParserSettings()
             {
-                return m;
+                NumberOfArrayItems = 2, // Important because coordinates are arrays with 2 items [lat, lon]
             }
         );
     }
@@ -47,5 +40,40 @@ public class GeoNorgeTestFixture : TestBedFixture
         _server.Stop();
         _server.Dispose();
         return ValueTask.CompletedTask;
+    }
+
+    private MappingModel KommuneFylkeMappingVisitor(MappingModel mapping)
+    {
+        mapping.Request.Path = mapping.Request.Path switch
+        {
+            string path and "/kommuneinfo/v1/fylker/example-string" => path.Replace(
+                "example-string",
+                "03"
+            ),
+            string path2 and "/kommuneinfo/v1/kommuner/example-string" => path2.Replace(
+                "example-string",
+                "0301"
+            ),
+            _ => mapping.Request.Path,
+        };
+
+        return mapping;
+    }
+}
+
+file static class Extensions
+{
+    public static void AddMappings(
+        this WireMockServer server,
+        string filePath,
+        Func<MappingModel, MappingModel>? mappingVisitor = null,
+        WireMockOpenApiParserSettings? settings = null
+    )
+    {
+        mappingVisitor ??= m => m;
+
+        using var fileStream = File.Open(filePath, FileMode.Open, FileAccess.Read, FileShare.Read);
+
+        server.AddOpenApiMappings(fileStream, mappingVisitor, settings);
     }
 }